+++ api/discovery/target.go
@@ -1,6 +1,7 @@
 package discovery
 
 import (
+	"context"
 	"sync"
 
 	"google.golang.org/grpc"
@@ -151,4 +152,31 @@ func (s *TargetObserverSet) notifyOne(
 	}
 
 	g.Wait()
+}
+
+// TargetExecutor is a TargetObserver that executes a function in a new
+// goroutine whenever a target becomes available.
+type TargetExecutor struct {
+	executor
+
+	// Func is the function to execute when a target becomes available.
+	// The context is canceled when the target becomes unavailable.
+	Func func(context.Context, *Target)
+
+	// Parent is the parent context under which the function is called.
+	// If it is nil, context.Background() is used.
+	Parent context.Context
+}
+
+// TargetAvailable starts a new goroutine for the given target.
+func (e *TargetExecutor) TargetAvailable(t *Target) {
+	e.start(e.Parent, t, func(ctx context.Context) {
+		e.Func(ctx, t)
+	})
+}
+
+// TargetUnavailable cancels the context associated with any existing goroutine
+// for the given target and waits for the goroutine to exit.
+func (e *TargetExecutor) TargetUnavailable(t *Target) {
+	e.stop(t)
 }