+++ CHANGELOG.md
@@ -11,9 +11,14 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Add `discovery.TargetExecutor` and `ClientExecutor`
+
 ### Changed
 
 - **[BC]** Change the internal gRPC API namespace from `dogma.configkit.v1` to `dogma.config.v1`
+- **[BC]** `discovery.Connector` no longer implements `TargetObserver`, instead use a `TargetExecutor` to call `Connector.Run()`
 
 ## [0.3.1] - 2020-03-08